refactor: move files into dirs
