+++ tonal_hortator/core/__init__.py
@@ -2,9 +2,9 @@
 Core functionality for Tonal Hortator
 """
 
-from .embeddings import OllamaEmbeddingService
-from .playlist_generator import LocalPlaylistGenerator
-from .track_embedder import LocalTrackEmbedder
+from .embeddings.embeddings import OllamaEmbeddingService
+from .embeddings.track_embedder import LocalTrackEmbedder
+from .playlist.playlist_generator import LocalPlaylistGenerator
 
 __all__ = [
     "OllamaEmbeddingService",
